+++ src/qtbox/dial/style/demo2.py
@@ -59,7 +59,7 @@ class QtBoxStyleDial2(QDial):
         painter.setFont(font)
         font_metrics = QFontMetricsF(font)
         text_width = font_metrics.width(str(self.value()))
-        painter.drawText(self.width() / 2 - text_width / 2, self.height() / 2, str(self.value()))
+        painter.drawText(int(self.width() / 2 - text_width / 2), int(self.height() / 2), str(self.value()))
 
     def get_handle_pos(self):
         center_x = self.rect().center().x()
@@ -165,7 +165,7 @@ class QtBoxStyleDial2(QDial):
 #         painter.setFont(font)
 #         font_metrics = QFontMetricsF(font)
 #         text_width = font_metrics.width(str(self.value()))
-#         painter.drawText(self.width() / 2 - text_width / 2, self.height() / 2, str(self.value()))
+#         painter.drawText(int(self.width() / 2 - text_width / 2), int(self.height() / 2), str(self.value()))
 
 #     def get_handle_pos(self):
 #         center_x = self.rect().center().x()
