+++ src/qtbox/progressbar/style/demo7.py
@@ -33,7 +33,7 @@ class QtBoxStyleProgressBar7(QProgressBar):
         painter.setFont(font)
         font_metrics = QFontMetricsF(font)
         text_width = font_metrics.width(self.text())
-        painter.drawText(self.width()/2-text_width/2, 45, self.text())
+        painter.drawText(int(self.width()/2-text_width/2), 45, self.text())
 # PyQt
 
 # PySide
@@ -68,5 +68,5 @@ class QtBoxStyleProgressBar7(QProgressBar):
 #         painter.setFont(font)
 #         font_metrics = QFontMetricsF(font)
 #         text_width = font_metrics.width(self.text())
-#         painter.drawText(self.width()/2-text_width/2, 45, self.text())
+#         painter.drawText(int(self.width()/2-text_width/2), 45, self.text())
 # PySide
